# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Alojamientos
  version: 1.0.0
  description: API para usuarios, anfitriones, alojamientos, reservas y comentarios. Rutas y respuestas basadas en los requerimientos del proyecto.
servers:
  - url: http://localhost:8080
paths:
...
  /reservas:
    post:
      summary: Crear una nueva reserva (usuario autenticado)
      operationId: crearReserva
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaCrear'
      responses:
        '201':
          description: Reserva creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservaResponse'
        '400':
          description: Datos inválidos (fechas, formato)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto (fechas ocupadas o capacidad excedida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Listar reservas (filtros) — uso administrativo / host
      operationId: listarReservas
      security:
        - BearerAuth: []
      parameters:
        - name: usuarioId
          in: query
          schema:
            type: string
        - name: alojamientoId
          in: query
          schema:
            type: string
        - name: estado
          in: query
          schema:
            type: string
            enum: [PENDIENTE, CONFIRMADA, CANCELADA, COMPLETADA]
        - name: fechaInicio
          in: query
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          schema:
            type: string
            format: date
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservaResponse'
        '400':
          description: Filtros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservas/{id}/cancelar:
    put:
      summary: Cancelar una reserva (usuario) — validar plazo 48h
      operationId: cancelarReserva
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reserva cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Cancelación no permitida (fuera de plazo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reserva no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reservas/{id}/comentarios:
    post:
      summary: Crear comentario (reseña) asociado a una reserva completada
      operationId: crearComentario
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComentarioCrear'
      responses:
        '201':
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComentarioResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Reserva no completada o no corresponde al usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Ya existe un comentario para esta reserva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comentarios/{id}/respuesta:
    post:
      summary: Responder un comentario (anfitrión)
      operationId: responderComentario
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespuestaCrear'
      responses:
        '201':
          description: Respuesta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: No es anfitrión o no es propietario del alojamiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Comentario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /alojamientos/{id}/comentarios:
    get:
      summary: Listar comentarios de un alojamiento
      operationId: listarComentariosAlojamiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: orden
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
        - name: limite
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de comentarios ordenados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComentarioResponse'
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: "Operación exitosa"
    ErrorResponse:
      type: object
      properties:
        codigo:
          type: string
          example: "ERR_VALIDACION"
        mensaje:
          type: string
          example: "El email ya está registrado"

    # -- Alojamiento --
...
    ReservaCrear:
      type: object
      required:
        - alojamientoId
        - checkIn
        - checkOut
        - huespedes
      properties:
        alojamientoId:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        huespedes:
          type: integer
    ReservaResponse:
      type: object
      properties:
        id:
          type: string
        alojamientoId:
          type: string
        usuarioId:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        huespedes:
          type: integer
        estado:
          type: string
          enum: [PENDIENTE, CONFIRMADA, CANCELADA, COMPLETADA]
        precioTotal:
          type: number

    # -- Comentarios y respuestas --
    ComentarioCrear:
      type: object
      required:
        - calificacion
      properties:
        calificacion:
          type: integer
          minimum: 1
          maximum: 5
        comentario:
          type: string
          maxLength: 500
    ComentarioResponse:
      type: object
      properties:
        id:
          type: string
        reservaId:
          type: string
        usuarioId:
          type: string
        calificacion:
          type: integer
        comentario:
          type: string
        creadoEn:
          type: string
          format: date-time
    RespuestaCrear:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
          maxLength: 500
    RespuestaResponse:
      type: object
      properties:
        id:
          type: string
        comentarioId:
          type: string
        anfitrionId:
          type: string
        mensaje:
          type: string
        creadoEn:
          type: string
          format: date-time